refactor(Avatar): file name match component name (#52593)
